+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project (loosely) adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+* Expand name characters to 6 bits wide and map all 64 characters
 
 ## [3.0.0](https://github.com/animavirtuality/onlybots-core/compare/2.1.0...3.0.0) - November 10, 2022
 ### Changed